% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPET_time.R
\docType{data}
\name{CPET_time}
\alias{CPET_time}
\title{SIPETool example files}
\description{
Raw and filtered output from SIFT-MS, and CPET system.
Four different files are available:
raw_sift -> raw data from SIFT-MS
CPET_time -> data from CPET system for time allignment
SIFT_time -> filtered data from SIFT-MS for time allignment
SIFT_filtered -> raw data from SIFT-MS filtered using SIFT_output_filter
}
\keyword{data}
