% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_alpha_0.06.R
\name{data_indexer}
\alias{data_indexer}
\title{data indexer}
\usage{
data_indexer(dat, time_frame_index = NA)
}
\arguments{
\item{dat}{input vector}

\item{time_frame_index}{custom data range from the time column}
}
\value{
a vector indexed according to the specified time frame
}
\description{
This function takes as input a vector and return the
data index according to the selected time frame
}
\examples{
data_indexer(c(1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1))

}
