% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_alpha_0.06.R
\name{normalizer}
\alias{normalizer}
\title{Data normalizer}
\usage{
normalizer(dat, norm_range = c(0, 1))
}
\arguments{
\item{dat}{the vector to normalize}

\item{norm_range}{the range used for normalization}
}
\value{
vector normalized between norm_range
}
\description{
This function takes as input a vector and returns it normalized between
a specified range
}
\examples{
normalizer(c(1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1))

}
