% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_SIFT.R
\docType{data}
\name{raw_SIFT}
\alias{raw_SIFT}
\title{SIPETool example files}
\description{
Raw and filtered output from SIFT-MS, and CPET system.
Four different files are available:
raw_SIFT -> raw data from SIFT-MS
CPET_time -> data from CPET system for time allignment
SIFT_time -> filtered data from SIFT-MS for time allignment
SIFT_filtered -> raw data from SIFT-MS filtered using SIFT_output_filter
}
\keyword{data}
