% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_alpha_0.06.R
\name{tidal_finder}
\alias{tidal_finder}
\title{Tidal finder}
\usage{
tidal_finder(
  dat,
  height_threshold = 0.2,
  refine = FALSE,
  time_frame_index = NA
)
}
\arguments{
\item{dat}{the input matrix}

\item{height_threshold}{the minimum height of the tidal}

\item{refine}{refine the dataset}

\item{time_frame_index}{custom time frame}
}
\value{
matrix with the tidals for each column
}
\description{
This function takes as input a matrix and returns for
each column the end tidals depending of the threshold set. It is
possible to set a custom time frame for the search of the tidals.
Note: a minimum amount of 45 points are necessary.
}
\examples{
tidal_finder(c(1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1))

}
