% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_allignment.R
\name{time_filter}
\alias{time_filter}
\title{Time_alignment}
\usage{
time_filter(Cy = file.choose(), sift = file.choose())
}
\arguments{
\item{Cy}{CPET-ESE output file}

\item{sift}{SIFT-MS refined file}
}
\value{
A plot and the SIFT-MS data file resized for the alignment with the CPET-ESE file
}
\description{
This function takes as input two data set containing a time vector and
a data vector and return the two data sets aligned. This is done
by reducing the dimensions of the data set with higher points. The first data set
is the one coming from the CPET-ESE and the second one from the SIFT-MS
}
\examples{
data(SIFT_time)
data(CPET_time)
time_filter(CPET_time, SIFT_time)

}
