% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_alpha_0.06.R
\name{trend}
\alias{trend}
\title{Trend finder}
\usage{
trend(dat)
}
\arguments{
\item{dat}{the vector to analyze}
}
\value{
vector containing the trend of the each column
}
\description{
This function takes as input a vector and returns
the trend of each column expressed as the difference
between two consecutive elements
}
\examples{
trend(c(1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1,1:10,10:1))

}
